requirements.txt
django==5.0.7
Pillow==10.4.0
faker==28.0.0